+++ sparkle/tools/__init__.py
@@ -4,4 +4,4 @@ from sparkle.tools.general import get_time_pid_random_string
 from sparkle.tools.parameters import PCSConverter
 from sparkle.tools.slurm_parsing import SlurmBatch
 from sparkle.tools.solver_wrapper_parsing import get_solver_call_params
-from sparkle.tools.runsolver import RunSolver
+from sparkle.tools.runsolver.runsolver import RunSolver